faler/requirements.txt
requests==2.12.4
future==0.16.0
python-telegram-bot
